feat: move extracted classes to package
