+++ app/src/main/java/com/machikoro/client/ui/game/ui/Marketplace.kt
@@ -1,4 +1,4 @@
-package com.machikoro.client.ui.game
+package com.machikoro.client.ui.game.ui
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
